requirements.txt
flask==2.3.2
flask-socketio==5.3.4
pyserial==3.5
numpy<2
matplotlib==3.8.4
colorama==0.4.6
python-socketio==5.8.0
eventlet==0.33.3